backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.9.0
django-axes==8.0.0
django-cachalot==2.8.0
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filer==3.3.1
django-filter==25.1
django-haystack==3.3.0
django-import-export==4.3.9
django-js-asset==3.1.2
django-payments==3.0.1
django-phonenumber-field==8.1.0
django-polymorphic==4.1.0
djangorestframework==3.16.0
drf-yasg==1.21.10
easy-thumbnails==2.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
lxml==5.4.0
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
phonenumberslite==9.0.6
pillow==11.2.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
xlrd==2.0.2
xlwt==1.3.0
